# Conflicts:
#	shop-web/src/main/java/com/masiis/shop/web/platform/controller/order/BOrderController.java

+++ shop-web/src/main/java/com/masiis/shop/web/platform/controller/order/BOrderController.java
@@ -184,10 +184,10 @@ public class BOrderController extends BaseController {
      */
     @RequestMapping("/payBOrder.shtml")
     public ModelAndView payBOrder(HttpServletRequest request,
-                                  HttpServletResponse response,
-                                  @RequestParam(value = "userAddressId", required = false) Integer userAddressId,
-                                  @RequestParam(value = "userMessage", required = false) String userMessage,
-                                  @RequestParam(value = "bOrderId", required = false) Long bOrderId
+                                     HttpServletResponse response,
+                                     @RequestParam(value = "userAddressId", required = false) Long userAddressId,
+                                     @RequestParam(value = "userMessage", required = false) String userMessage,
+                                     @RequestParam(value = "bOrderId", required = false) Long bOrderId
     ) {
         ModelAndView mv = new ModelAndView();
         String skuImg = PropertiesUtils.getStringValue("index_product_220_220_url");
